Added data analysis files
These two files assist with analyzing the data that is stored in the log buckets in Google Cloud.
`parse.json` parses the log files to extract the pertinent data and save this to a separate log file, one per input file
`analyze_logs.ipynb` is a Jupyter notebook that analyzes these files to produce a scatter plot of when the chat tool was used. It is reliant on a CSV file that contains the date of each design challenge session as well as the log files.